requirements.txt
hikari
hikari-crescent
python-dotenv
hikari-miru
psycopg2-binary
asyncpg